test: relocate audio private unit tests
