+++ docs/specs/effect-adoption-roadmap.html
@@ -1389,7 +1389,7 @@
 						<div class="metric-list">
 							<div class="metric">
 								<span class="badge good">Known</span>
-								<div>ABB current docs define the product spine, layer model, boundary rule, and five grey-box public APIs.</div>
+								<div>ABB current docs define the product spine, layer model, boundary rule, and six grey-box public APIs.</div>
 							</div>
 							<div class="metric">
 								<span class="badge good">Known</span>
